BATM-1197 refactor - remove LBTC and DAI PaymentSupport duplicates
This commit does not change any functionality it's just a refactor to be able
to use the common code in both LBTC and DAI payment supports in SLP payment support
