+++ server_extensions_extra/src/main/java/com/generalbytes/batm/server/extensions/extra/common/PollingPaymentSupport.java
@@ -1,0 +1,135 @@
+/*************************************************************************************
+ * Copyright (C) 2014-2019 GENERAL BYTES s.r.o. All rights reserved.
+ *
+ * This software may be distributed and modified under the terms of the GNU
+ * General Public License version 2 (GPL2) as published by the Free Software
+ * Foundation and appearing in the file GPL2.TXT included in the packaging of
+ * this file. Please note that GPL2 Section 2[b] requires that all works based
+ * on this software must also be made publicly available under the terms of
+ * the GPL2 ("Copyleft").
+ *
+ * Contact information
+ * -------------------
+ *
+ * GENERAL BYTES s.r.o.
+ * Web      :  http://www.generalbytes.com
+ *
+ ************************************************************************************/
+package com.generalbytes.batm.server.extensions.extra.common;
+
+import com.generalbytes.batm.server.extensions.IExtensionContext;
+import com.generalbytes.batm.server.extensions.payment.IPaymentRequestListener;
+import com.generalbytes.batm.server.extensions.payment.IPaymentRequestSpecification;
+import com.generalbytes.batm.server.extensions.payment.IPaymentSupport;
+import com.generalbytes.batm.server.extensions.payment.PaymentReceipt;
+import com.generalbytes.batm.server.extensions.payment.PaymentRequest;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.math.BigDecimal;
+import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.Executors;
+import java.util.concurrent.ScheduledExecutorService;
+import java.util.concurrent.ScheduledFuture;
+import java.util.concurrent.TimeUnit;
+
+public abstract class PollingPaymentSupport implements IPaymentSupport {
+    protected static final Logger log = LoggerFactory.getLogger(PollingPaymentSupport.class);
+
+    protected final Map<String, PaymentRequest> requests = new ConcurrentHashMap<>();
+    protected final ScheduledExecutorService executorService = Executors.newSingleThreadScheduledExecutor();
+
+    @Override
+    public boolean init(IExtensionContext context) {
+        return true;
+    }
+
+    @Override
+    public PaymentRequest createPaymentRequest(IPaymentRequestSpecification spec) {
+        if (spec.getOutputs().size() != 1) {
+            throw new IllegalStateException("Only 1 output supported");
+        }
+        String address = spec.getOutputs().get(0).getAddress();
+
+        long validTillMillis = System.currentTimeMillis() + (spec.getValidInSeconds() * 1000);
+
+        PaymentRequest request = new PaymentRequest(spec.getCryptoCurrency(), spec.getDescription(), validTillMillis,
+            address, spec.getTotal(), BigDecimal.ZERO, spec.getRemoveAfterNumberOfConfirmationsOfIncomingTransaction(), spec.getRemoveAfterNumberOfConfirmationsOfOutgoingTransaction(), spec.getWallet(), spec.getTimeoutRefundAddress(),
+            spec.getOutputs(), spec.isDoNotForward(), null);
+
+        registerPaymentRequest(request);
+        return request;
+    }
+
+
+    @Override
+    public void registerPaymentRequest(PaymentRequest request) {
+
+        ScheduledFuture<?> scheduledFuture = executorService.scheduleAtFixedRate(() -> poll(request), getPollingInitialDelayMillis(), getPollingPeriodMillis(), TimeUnit.MILLISECONDS);
+
+        executorService.schedule(() -> {
+            try {
+                scheduledFuture.cancel(false);
+                if (request.getState() != PaymentRequest.STATE_SEEN_IN_BLOCK_CHAIN) {
+                    log.info("Cancelling {}", request);
+                    setState(request, PaymentRequest.STATE_TRANSACTION_TIMED_OUT);
+                }
+            } catch (Throwable t) {
+                log.error("", t);
+            }
+        }, request.getValidTill() - System.currentTimeMillis(), TimeUnit.MILLISECONDS);
+
+        requests.entrySet().removeIf(e -> e.getValue().getValidTill() < System.currentTimeMillis());
+        requests.put(request.getAddress(), request);
+    }
+
+    @Override
+    public boolean isPaymentReceived(String paymentAddress) {
+        PaymentRequest paymentRequest = requests.get(paymentAddress);
+        return paymentRequest != null && paymentRequest.getState() == PaymentRequest.STATE_SEEN_IN_BLOCK_CHAIN;
+    }
+
+    @Override
+    public PaymentReceipt getPaymentReceipt(String paymentAddress) {
+        PaymentReceipt result = new PaymentReceipt(getCryptoCurrency(), paymentAddress);
+        PaymentRequest paymentRequest = requests.get(paymentAddress);
+        if (paymentRequest != null && paymentRequest.getState() == PaymentRequest.STATE_SEEN_IN_BLOCK_CHAIN) {
+            result.setStatus(PaymentReceipt.STATUS_PAID);
+            result.setConfidence(PaymentReceipt.CONFIDENCE_SURE);
+            result.setAmount(paymentRequest.getAmount());
+            result.setTransactionId(paymentRequest.getIncomingTransactionHash());
+        }
+        return result;
+    }
+
+    protected void fireNumberOfConfirmationsChanged(PaymentRequest request, int numberOfConfirmations) {
+        IPaymentRequestListener listener = request.getListener();
+        if (listener != null) {
+            listener.numberOfConfirmationsChanged(request, numberOfConfirmations, IPaymentRequestListener.Direction.INCOMING);
+        }
+    }
+
+    protected void setState(PaymentRequest request, int newState) {
+        int previousState = request.getState();
+        request.setState(newState);
+        log.debug("Transaction state changed: {} -> {} {}", previousState, newState, request);
+
+        IPaymentRequestListener listener = request.getListener();
+        if (listener != null) {
+            listener.stateChanged(request, previousState, request.getState());
+        }
+    }
+
+    protected long getPollingPeriodMillis() {
+        return TimeUnit.SECONDS.toMillis(5);
+    }
+
+    protected long getPollingInitialDelayMillis() {
+        return TimeUnit.SECONDS.toMillis(15);
+    }
+
+    protected abstract String getCryptoCurrency();
+
+    protected abstract void poll(PaymentRequest request);
+}
